Update README.md
Fleshed out the overall specs of the workshop, how to set it up, how to run it (in a general sense), and provided a method of contact if anyone wants to improve the club. Also, added a suggestion for how to use what was learned in this workshop for a future club meeting idea (competition).